+++ src/main/java/com/ebicep/warlords/pve/items/statpool/ItemGauntletStatPool.java
@@ -7,34 +7,34 @@ public enum ItemGauntletStatPool implements ItemStatPool<ItemGauntletStatPool> {
 
     HP("Health") {
         @Override
-        public void applyToWarlordsPlayer(WarlordsPlayer warlordsPlayer, float value) {
+        public void applyToWarlordsPlayer(WarlordsPlayer warlordsPlayer, int value) {
             warlordsPlayer.setMaxBaseHealth(warlordsPlayer.getMaxBaseHealth() + value);
         }
     },
     MAX_ENERGY("Max Energy") {
         @Override
-        public void applyToWarlordsPlayer(WarlordsPlayer warlordsPlayer, float value) {
+        public void applyToWarlordsPlayer(WarlordsPlayer warlordsPlayer, int value) {
             AbstractPlayerClass playerClass = warlordsPlayer.getSpec();
-            playerClass.setMaxEnergy((int) (playerClass.getMaxEnergy() + value));
+            playerClass.setMaxEnergy(playerClass.getMaxEnergy() + value);
         }
     },
     EPH("Energy Per Hit") {
         @Override
-        public void applyToWarlordsPlayer(WarlordsPlayer warlordsPlayer, float value) {
+        public void applyToWarlordsPlayer(WarlordsPlayer warlordsPlayer, int value) {
             AbstractPlayerClass playerClass = warlordsPlayer.getSpec();
             playerClass.setEnergyOnHit(playerClass.getEnergyOnHit() + value);
         }
     },
     EPS("Energy Per Second") {
         @Override
-        public void applyToWarlordsPlayer(WarlordsPlayer warlordsPlayer, float value) {
+        public void applyToWarlordsPlayer(WarlordsPlayer warlordsPlayer, int value) {
             AbstractPlayerClass playerClass = warlordsPlayer.getSpec();
             playerClass.setEnergyPerSec(playerClass.getEnergyPerSec() + value);
         }
     },
     SPEED("Speed") {
         @Override
-        public void applyToWarlordsPlayer(WarlordsPlayer warlordsPlayer, float value) {
+        public void applyToWarlordsPlayer(WarlordsPlayer warlordsPlayer, int value) {
             warlordsPlayer.getSpeed().addBaseModifier(value);
         }
 
@@ -68,5 +68,5 @@ public enum ItemGauntletStatPool implements ItemStatPool<ItemGauntletStatPool> {
         return Operation.ADD;
     }
 
-    public abstract void applyToWarlordsPlayer(WarlordsPlayer warlordsPlayer, float value);
+    public abstract void applyToWarlordsPlayer(WarlordsPlayer warlordsPlayer, int value);
 }
